Abstract dashboard. Add policy for home controller.
