+++ app/controllers/dashboard_controller.rb
@@ -1,0 +1,7 @@
+class DashboardController < ApplicationController
+  def index
+    authorize Current.user, policy_class: HomePolicy
+
+    @organization = Organization.find(Current.user.current_organization_id)
+  end
+end
