+++ app/controllers/home_controller.rb
@@ -1,13 +1,9 @@
 class HomeController < ApplicationController
   def index
+    authorize Current.session
+
     if Current.user.current_organization_id.present?
       redirect_to dashboard_path
     end
-  end
-
-  def dashboard
-    authorize Current.user, policy_class: HomePolicy
-
-    @organization = Organization.find(Current.user.current_organization_id)
   end
 end
