+++ app/helpers/dashboard_helper.rb
@@ -1,0 +1,2 @@
+module DashboardHelper
+end
