+++ app/policies/dashboard_policy.rb
@@ -1,0 +1,24 @@
+class DashboardPolicy < ApplicationPolicy
+  # NOTE: Up to Pundit v2.3.1, the inheritance was declared as
+  # `Scope < Scope` rather than `Scope < ApplicationPolicy::Scope`.
+  # In most cases the behavior will be identical, but if updating existing
+  # code, beware of possible changes to the ancestors:
+  # https://gist.github.com/Burgestrand/4b4bc22f31c8a95c425fc0e30d7ef1f5
+
+  class Scope < ApplicationPolicy::Scope
+    def index?
+      return false unless user.current_organization.present?
+
+      can_access_dashboard?
+    end
+
+    private
+    def organization
+      user.current_organization
+    end
+
+    def can_access_dashboard?
+      organization.accounts.include?(user)
+    end
+  end
+end
