+++ app/policies/home_policy.rb
@@ -5,18 +5,7 @@ class HomePolicy < ApplicationPolicy
   # code, beware of possible changes to the ancestors:
   # https://gist.github.com/Burgestrand/4b4bc22f31c8a95c425fc0e30d7ef1f5
 
-  def dashboard?
-    return false unless user.current_organization.present?
-
-    can_access_dashboard?
-  end
-
-  private
-  def organization
-    user.current_organization
-  end
-
-  def can_access_dashboard?
-    organization.accounts.include?(user)
+  def index?
+    record.present?
   end
 end
